+++ lib/src/bindings/soloud_controller.dart
@@ -1,8 +1,8 @@
 import 'dart:ffi' as ffi;
 import 'dart:io';
 
-import 'package:flutter_soloud/src/bindings_capture_ffi.dart';
-import 'package:flutter_soloud/src/bindings_player_ffi.dart';
+import 'package:flutter_soloud/src/bindings/bindings_capture_ffi.dart';
+import 'package:flutter_soloud/src/bindings/bindings_player_ffi.dart';
 
 /// Controller that expose method channel and FFI
 class SoLoudController {
